Adicionando CRUD completo das classes empresa, genero e console. Subindo também a implementação do cálculo da nota dentro de AvaliacaoService

+++ src/main/java/br/edu/ufabc/rankingJogos/src/core/port/repository/GeneroRepositoryPort.java
@@ -12,11 +12,11 @@ public interface GeneroRepositoryPort {
 
     int update(Genero genero);
 
-    int deleteById(Long id);
+    int deleteById(int id);
 
     List<Genero> findAll();
 
-    Genero findById(Long id);
+    Genero findById(int id);
     
     List<Genero> findByGenero(String genero);
 }
