+++ src/main/java/br/edu/ufabc/rankingJogos/src/core/service/AvaliacaoService.java
@@ -38,7 +38,7 @@ public class AvaliacaoService {
         return avaliacaoRepository.delete(id);
     }
     
-    public int calculaNota(int jogo_id){
+    public Double calculaNota(int jogo_id){
         return avaliacaoRepository.nota(jogo_id);
     }
 }
